feat(demo): modularize + test the demo command, retire shell scripts
